Refactoring code

+++ src/main/java/com/sap/service/impl/UserServiceImp.java
@@ -2,6 +2,7 @@ package com.sap.service.impl;
 import com.sap.dao.UserDao;
 import com.sap.service.UserService;
 import com.sap.models.User;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 import javax.annotation.Resource;
 import java.util.List;
@@ -21,6 +22,7 @@ public class UserServiceImp implements UserService {
 
     @Override
     public void updateUser(User user) {
+        if(notAuthorized(user.getId())){ throw new RuntimeException("You are not allowed to execute this action, you idiot."); }
         this.userDao.updateUser(user);
     }
 
@@ -36,6 +38,27 @@ public class UserServiceImp implements UserService {
 
     @Override
     public void removeUser(int id) {
+        if(notAuthorized(id)){ throw new RuntimeException(); }
         this.userDao.removeUser(id);
     }
+
+    /**
+     * Verify if user is not trying to execute an action at himself or at some member that is not from his team.
+     * @param id
+     * @return Boolean
+     */
+    public Boolean notAuthorized(int id) {
+        User principal = this.getUserByName(SecurityContextHolder.getContext().getAuthentication().getName());
+        // User cannot execute action at himself !IMPORTANT
+        if (principal.getId() == id) {
+            return true;
+        }
+        User user = this.getUserById(id);
+        // Owner cannot execute an action with a member that is not from his team !IMPORTANT
+        if (principal.getTeam().getId() != user.getTeam().getId()) {
+            return true;
+        }
+        return false;
+    }
+
 }
